added logo-mark.jsx to be shared and used elsewhere
