+++ src/components/modals/about-iodide.jsx
@@ -3,50 +3,11 @@
 import React from "react";
 import styled from "react-emotion";
 import HelpModalContent from "./help-modal-content";
+import LogoMark from "../../shared/components/logo/logo-mark";
 import ThreeElements from "../../shared/components/three-set/three-elements";
 import Element, {
   ElementTitle
 } from "../../shared/components/three-set/element";
-
-const IodideLogoMarkContainer = styled("svg")`
-  margin: auto;
-  display: block;
-`;
-const IodideLogoMark = () => (
-  <IodideLogoMarkContainer
-    width="100"
-    viewBox="0 0 386 386"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <rect
-      x="15"
-      y="15"
-      width="356"
-      height="356"
-      rx="30"
-      fill="white"
-      stroke="#4f3554"
-      strokeWidth="30"
-      strokeMiterlimit="1"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-    <path
-      d="M240.6 326H141.4V294.4H173V97.6H141.4V66H240.6V97.6H209.4V294.4H240.6V326Z"
-      fill="#4f3554"
-    />
-    <line
-      x1="265"
-      y1="112.5"
-      x2="329"
-      y2="112.5"
-      stroke="#4f3554"
-      strokeWidth="27"
-      strokeLinejoin="round"
-    />
-  </IodideLogoMarkContainer>
-);
 
 const AboutContentContainer = styled("div")`
   width: 700px;
@@ -117,7 +78,7 @@ const MiddleSection = styled("div")`
 export default () => (
   <HelpModalContent>
     <AboutContentContainer>
-      <IodideLogoMark />
+      <LogoMark />
       <AboutContentHeader>
         <IodideTitle>
           Iodide<sub>α</sub> <IodideVersion>{IODIDE_VERSION}</IodideVersion>
